requirements.txt
aiohttp
pyyaml==5.4.1
discord.py==1.6.0
youtube-dl==2021.3.2
PyNaCl==1.4.0
pycparser==2.20
async-timeout==3.0.1
regex==2020.11.13
PyYAML==5.4.1
